chore: Remove SizeMe. Move, delete utilities
