+++ src/plugins/graph/axis/components/axis-drag-rects.tsx
@@ -2,7 +2,7 @@ import {observer} from "mobx-react-lite";
 import React, {useEffect, useRef} from "react";
 import {reaction} from "mobx";
 import {drag, ScaleContinuousNumeric, select, Selection} from "d3";
-import t from "../../../../utilities/translation/translate";
+import t from "../../utilities/translation/translate";
 import {useAxisLayoutContext} from "../models/axis-layout-context";
 import {INumericAxisModel} from "../models/axis-model";
 import {isVertical} from "../../axis-graph-shared";
