+++ src/plugins/graph/utilities/math-utils.ts
