+++ src/plugins/graph/utilities/translation/translate.ts
